Rename test and make a Webform CiviCRM test bases
