+++ tests/src/FunctionalJavascript/WebformCivicrmTestBase.php
@@ -1,0 +1,66 @@
+<?php
+
+namespace Drupal\Tests\webform_civicrm\FunctionalJavascript;
+
+use Drupal\Tests\webform\Traits\WebformBrowserTestTrait;
+
+abstract class WebformCivicrmTestBase extends CiviCrmTestBase {
+
+  use WebformBrowserTestTrait;
+
+  /**
+   * {@inheritdoc}
+   */
+  protected static $modules = [
+    'webform',
+    'webform_ui',
+    'webform_civicrm',
+  ];
+
+  /**
+   * {@inheritdoc}
+   *
+   * During tests configuration schema is validated. This module does not
+   * provide schema definitions for its handler.
+   *
+   * To fix: webform.webform.civicrm_webform_test:handlers.webform_civicrm.settings
+   *
+   * @see \Drupal\Core\Test\TestSetupTrait::getConfigSchemaExclusions
+   */
+  protected static $configSchemaCheckerExclusions = [
+    'webform.webform.civicrm_webform_test',
+  ];
+
+  /**
+   * The test webform.
+   *
+   * @var \Drupal\webform\WebformInterface
+   */
+  protected $webform;
+
+  /**
+   * The test user.
+   *
+   * @var \Drupal\user\Entity\User
+   */
+  protected $testUser;
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function setUp() {
+    parent::setUp();
+    $this->testUser = $this->createUser([
+      'administer CiviCRM',
+      'access CiviCRM',
+      'access administration pages',
+      'access webform overview',
+      'administer webform',
+    ]);
+    $this->webform = $this->createWebform([
+      'id' => 'civicrm_webform_test',
+      'title' => 'CiviCRM Webform Test',
+    ]);
+  }
+
+}
